+++ src/server/api/logic/session.ts
@@ -1,21 +1,15 @@
 import { UserProfile, UserSubType, UserType } from "../../../types/user";
-export class UserTypeLogic {
-  userType: UserType;
-
-  constructor(userType: UserType) {
-    this.userType = userType;
-  }
-
-  isPetOwner() {
-    const isPetOwner: boolean = [UserType.PetOwner].includes(this.userType);
+export abstract class UserTypeLogic {
+  public static isPetOwner(userType: UserType) {
+    const isPetOwner: boolean = [UserType.PetOwner].includes(userType);
     return isPetOwner;
   }
 
-  isPetSitter() {
+  public static isPetSitter(userType: UserType) {
     const isPetSitter: boolean = [
       UserType.FreelancePetSitter,
       UserType.PetHotel,
-    ].includes(this.userType);
+    ].includes(userType);
     return isPetSitter;
   }
 }
