+++ src/server/api/routers/petSitterRouter.ts
@@ -11,7 +11,7 @@ import {
   userFields,
   searchField,
 } from "../../../schema/schema";
-import * as searchLogic from "../logic/search";
+import { petSitterSearchLogic } from "../logic/search/petSitter";
 
 const zodUserFields = z.object({
   verifyStatus: z.boolean(),
@@ -108,50 +108,6 @@ export const petSitterRouter = createTRPCRouter({
       });
       return update;
     }),
-  // searchPetSitter
-  // input:
-  // note that all args default value will act as those fields are not in search criteria
-  // read more about default value on searchField
-  //   searchName: @string name such as "john smith"
-  //   searchRating: @number not implement yet
-  //   searchPriceMin:  @number minimum price that user is ok with
-  // searchPriceMin can be null and default is null   ( also endPrice )
-  // table below show when petsitter will be in search result (✓ mean show in result)
-  // +---------------+------+---------------+
-  // | input  \   db | null | num(endPrice) |
-  // +---------------+------+---------------+
-  // | null          | ✓    | ✓            |
-  // +---------------+------+---------------+
-  // | num(priceMin) | x    |  if endPrice  |
-  // |               |      |  >  priceMin  |
-  // +---------------+------+---------------+
-  //   searchPriceMax:  @number maximum price that user is ok with
-  //   searchLocation:  @string not implement yet
-  //   searchPetType:   @string pet types such as "cat dog" it will return petsitter that have both cat and dog
-  //   searchStartSchedule: @string not implement yet
-  //   searchEndSchedule:   @string not implement yet
-  //   searchIncludePetHotelFlag: @boolean if you want to include pet hotel in search result then set it to true
-  //   searchIncludeFreelancePetSitterFlag: @boolean if you want to inclde FreelancePetSitter in search result then set it to true
-  // table to show how to set value of searchIncludePetHotelFlag and searchIncludeFreelancePetSitterFlag
-  // +-------------------------------+---------------------------+-------------------------------------+
-  // | want to seach for \ args name | searchIncludePetHotelFlag | searchIncludeFreelancePetSitterFlag |
-  // +-------------------------------+---------------------------+-------------------------------------+
-  // | pet hotel                     | true                      | false                               |
-  // +-------------------------------+---------------------------+-------------------------------------+
-  // | freelance                     | false                     | true                                |
-  // +-------------------------------+---------------------------+-------------------------------------+
-  // | all                           | true                      | true                                |
-  // +-------------------------------+---------------------------+-------------------------------------+
-  //  searchSortBy: @string for sorting. support "name", "username", "price", "startPrice", "endPrice" default is username
-  //  will implement sort by firstname lastname hotel name if needed
-
-  // output:
-  //   list of object that only contain user id that match search input
-  //   example:
-  //   [
-  //     { "userId": "cldzsm2n6000z7k083zwda6b1" },
-  //     { "userId": "cldzsm3tt00137k08lt9r3rfs" },
-  //   ]
   searchPetSitter: publicProcedure
     .input(searchField)
     .query(async ({ ctx, input }) => {
@@ -161,28 +117,33 @@ export const petSitterRouter = createTRPCRouter({
         cursor: input.cursor ? { userId: input.cursor } : undefined,
         where: {
           AND: [
-            input.searchName ? searchLogic.searchByName(input.searchName) : {},
+            input.searchName
+              ? petSitterSearchLogic.petSitterName(input.searchName)
+              : {},
             input.searchPriceMin
-              ? searchLogic.searchByPriceMin(input.searchPriceMin)
+              ? petSitterSearchLogic.priceMin(input.searchPriceMin)
               : {},
             input.searchPriceMax
-              ? searchLogic.searchByPriceMax(input.searchPriceMax)
+              ? petSitterSearchLogic.priceMax(input.searchPriceMax)
               : {},
             input.searchPetTypes
-              ? searchLogic.searchByPetTypes(input.searchPetTypes)
+              ? petSitterSearchLogic.petTypes(input.searchPetTypes)
+              : {},
+            input.searchVerifyStatus
+              ? petSitterSearchLogic.verifyStatus(input.searchVerifyStatus)
               : {},
             input.searchIncludePetSitterType
-              ? searchLogic.searchByPetSitterTypes(
+              ? petSitterSearchLogic.petSitterTypes(
                   input.searchIncludePetSitterType
                 )
               : {},
-            searchLogic.searchByPetSitterType(
+            petSitterSearchLogic.petSitterType(
               input.searchIncludePetHotel,
               input.searchIncludeFreelancePetSitter
             ),
           ],
         },
-        orderBy: [searchLogic.searchOrderBy(input.searchSortBy)],
+        orderBy: [petSitterSearchLogic.sortBy(input.searchSortBy)],
         include: {
           user: true,
           petHotel: true,
